rental_manager/requirements.txt
flask
flask-login
flask-sqlalchemy
fpdf